refactor: joinArticlesを別のファイルに移動
