+++ utils/articles.ts
@@ -2,8 +2,6 @@ import { readFile } from "fs/promises"
 import type { ArticleMetadata } from "article"
 import { glob } from "glob"
 import matter from "gray-matter"
-import type { fetchBookmarksByUsername } from "@/actions/bookmark-actions"
-import type { fetchArticlesByUsername} from "@/actions/like-actions";
 import { fetchLikeCount } from "@/actions/like-actions"
 
 export const getArticlePaths = async (dir: string) => {
@@ -55,18 +53,4 @@ export const getArticleContent = async (
     console.error(`Failed to read markdown file: ${error}`)
     return { content: "" }
   }
-}
-
-export function joinArticles(
-  fetchedArticles: Awaited<ReturnType<typeof fetchArticlesByUsername>> | Awaited<ReturnType<typeof fetchBookmarksByUsername>>,
-  articles: Awaited<ReturnType<typeof getArticleList>>
-) {
-  return articles.map((article) => {
-    const isHit = fetchedArticles.some(
-      (item) => item.articleURL === article.slug,
-    )
-    return isHit ? article : null // 一致した場合は記事を返し、一致しない場合はnull
-  })
-    // nullを除外して最終的なlikedArticlesを得る
-    .filter((article) => article !== null)
 }
